+++ sdk/io/basic/payload.hpp
@@ -14,8 +14,8 @@
 #include <functional>
 #include <list>
 #include <map>
-#include <sdk/base/basic/template.hpp>
 #include <sdk/base/basic/variant.hpp>
+#include <sdk/base/nostd/integer.hpp>
 #include <sdk/base/system/endian.hpp>
 #include <sdk/base/system/shared_instance.hpp>
 #include <sdk/base/template.hpp>