+++ sdk/net/types.hpp
@@ -35,6 +35,7 @@
 
 #include <sdk/base/charset.hpp>
 #include <sdk/base/error.hpp>
+#include <sdk/base/nostd/range.hpp>
 #include <sdk/base/syntax.hpp>
 #include <sdk/base/types.hpp>
 #include <sdk/crypto/basic/types.hpp>